@@ -43,7 +43,7 @@
   "homepage": "https://github.com/scintilla-engine/scintilla-ide#readme",
   "devDependencies": {
     "@k-foss/ts-esnode": "^1.6.0",
-    "@types/copy-webpack-plugin": "^5.0.0",
+    "@types/copy-webpack-plugin": "^6.0.0",
     "@types/debounce-promise": "^3.1.1",
     "@types/express": "^4.17.6",
     "@types/fs-extra": "^9.0.1",
@@ -55,28 +55,28 @@
     "@types/node-ipc": "^9.1.3",
     "@types/proper-lockfile": "^3.0.1",
     "@types/stack-trace": "0.0.29",
-    "@types/webpack": "^4.41.17",
+    "@types/webpack": "^4.41.21",
     "@types/webpack-dev-middleware": "^3.7.1",
     "@types/webpack-env": "^1.15.2",
     "@types/webpack-hot-middleware": "^2.25.3",
     "@types/webpack-merge": "^4.1.5",
     "@types/webpack-notifier": "^1.5.3",
     "@types/yargs": "^15.0.5",
-    "@typescript-eslint/eslint-plugin": "^3.2.0",
-    "@typescript-eslint/parser": "^3.2.0",
+    "@typescript-eslint/eslint-plugin": "^3.10.1",
+    "@typescript-eslint/parser": "^3.10.1",
     "async-exit-hook": "^2.0.1",
     "awesome-typescript-loader": "^5.2.0",
     "chalk": "^4.1.0",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.1.1",
+    "copy-webpack-plugin": "^6.0.4",
     "crocket": "^0.9.11",
     "cross-env": "^6.0.3",
     "css-loader": "^3.4.1",
     "electron-devtools-installer": "^3.0.0",
-    "eslint": "^6.8.0",
+    "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-loader": "^3.0.3",
+    "eslint-loader": "^4.0.2",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-prettier": "^3.1.3",
     "express": "^4.17.1",
@@ -94,22 +94,22 @@
     "resolve-url-loader": "^3.1.1",
     "style-loader": "^1.0.1",
     "svg-inline-loader": "^0.8.0",
-    "ts-loader": "^7.0.5",
-    "ts-node": "^8.10.2",
+    "ts-loader": "^8.0.3",
+    "ts-node": "^9.0.0",
     "tsconfig-paths": "^3.9.0",
-    "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "typescript": "^3.9.5",
+    "tsconfig-paths-webpack-plugin": "^3.3.0",
+    "typescript": "^4.0.2",
     "url-loader": "^3.0.0",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12",
     "webpack-dev-middleware": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0",
-    "webpack-merge": "^4.2.2",
+    "webpack-merge": "^5.1.2",
     "webpack-notifier": "^1.8.0"
   },
   "dependencies": {
     "@phosphor/widgets": "^1.9.3",
-    "chokidar": "^3.4.0",
+    "chokidar": "^3.4.2",
     "debounce-promise": "^3.1.2",
     "debug-menu": "^0.6.1",
     "document-register-element": "^1.14.3",
